made it so remote recipes are stored less often and we can now add them directly to a meal

+++ app/src/main/java/org/cook_e/cook_e/EditRecipeActivity.java
@@ -276,9 +276,12 @@ public class EditRecipeActivity extends AppCompatActivity
     private void deleteRecipe() {
         try {
             App.getAccessor().deleteRecipe(mRecipe);
+
             String imageLink = mRecipe.getImageLink();
-            File imageFile = new File(imageLink);
-            imageFile.delete();
+            if (imageLink != null) {
+                File imageFile = new File(imageLink);
+                imageFile.delete();
+            }
             finish();
         } catch (SQLException e) {
             new AlertDialog.Builder(this)
